Add VHeaderDesktop and remove unused components
