+++ src/layouts/default.vue
@@ -5,7 +5,10 @@
       <VMigrationNotice />
       <VTranslationStatusBanner />
       <template v-if="isNewHeaderEnabled">
-        <VHeader v-if="isSearchRoute" />
+        <template v-if="isSearchRoute">
+          <VHeaderDesktop v-if="isMinScreenLg" />
+          <VHeader v-else />
+        </template>
         <VHeaderInternal v-else />
       </template>
       <VHeaderOld v-else />
@@ -64,6 +67,7 @@ const embeddedPage = {
     VMigrationNotice,
     VTranslationStatusBanner,
     VHeaderOld,
+    VHeaderDesktop: () => import('~/components/VHeader/VHeaderDesktop.vue'),
     VHeaderInternal: () => import('~/components/VHeader/VHeaderInternal.vue'),
     VHeader: () => import('~/components/VHeader/VHeader.vue'),
     VFooter,
@@ -117,6 +121,7 @@ const embeddedPage = {
     return {
       isHeaderScrolled,
       isMinScreenMd,
+      isMinScreenLg,
       isSidebarVisible,
       isSearchRoute,
       headerHasTwoRows,
